chore(doprocess): Added embedded option to document and process run
