+++ web/apps/doprocess.app/components/processes/documents/DocumentDetailsToolbar.tsx
@@ -2,11 +2,12 @@
 
 import { useState } from 'react';
 import { cx } from 'classix';
-import { Delete, MoreHorizontal } from '@signalco/ui-icons';
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@signalco/ui/dist/Menu';
+import { Delete, Embed, MoreHorizontal } from '@signalco/ui-icons';
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuSeparator, DropdownMenuTrigger } from '@signalco/ui/dist/Menu';
 import { IconButton } from '@signalco/ui/dist/IconButton';
 import { Toolbar } from '../../shared/Toolbar';
 import { SavingIndicator } from '../../shared/SavingIndicator';
+import { EmbedModal } from '../../shared/EmbedModal';
 import { KnownPages } from '../../../src/knownPages';
 import { useDocument } from '../../../src/hooks/useDocument';
 import { DocumentDeleteModal } from './DocumentDeleteModal';
@@ -19,6 +20,7 @@ type DocumentDetailsToolbarProps = {
 export function DocumentDetailsToolbar({ id, saving }: DocumentDetailsToolbarProps) {
     const { data: document } = useDocument(id);
     const [deleteModalOpen, setDeleteModalOpen] = useState(false);
+    const [embedOpen, setEmbedOpen] = useState(false);
 
     return (
         <>
@@ -35,6 +37,10 @@ export function DocumentDetailsToolbar({ id, saving }: DocumentDetailsToolbarPro
                         </IconButton>
                     </DropdownMenuTrigger>
                     <DropdownMenuContent>
+                        <DropdownMenuItem startDecorator={<Embed />} onClick={() => setEmbedOpen(true)}>
+                                    Embed...
+                        </DropdownMenuItem>
+                        <DropdownMenuSeparator />
                         <DropdownMenuItem startDecorator={<Delete />} onClick={() => setDeleteModalOpen(true)}>
                             Delete...
                         </DropdownMenuItem>
@@ -48,6 +54,12 @@ export function DocumentDetailsToolbar({ id, saving }: DocumentDetailsToolbarPro
                     onOpenChange={setDeleteModalOpen}
                     redirect={KnownPages.Documents} />
             )}
+            <EmbedModal
+                header="Embed document"
+                subHeader="To embed this process run, copy the following HTML snippet and paste it into your website:"
+                src={`${window.location.origin}${KnownPages.Document(id)}/embedded`}
+                open={embedOpen}
+                onOpenChange={setEmbedOpen} />
         </>
     );
 }
